check against current regions to avoid overlaps
